+++ repro/models/durmus2020.py
@@ -1,0 +1,2 @@
+from ._models.durmus2020.src.model import FEQA
+from ._models.durmus2020.src.setup import Durmus2020SetupSubcommand